chore: refactor bundle module
