+++ packages/bundler-builder/package.json
@@ -11,7 +11,8 @@
     "build": "npx tsc -p ./tsconfig.json && cp ../shared/sim/tracer.js dist/shared/sim",
     "dev": "nodemon",
     "start": "node --import ./register-hook.js ./src/index.ts --txMode base | pino-pretty",
-    "start:prod": "echo no start:prod",
+    "start:unsafe": "node --import ./register-hook.js ./src/index.ts --txMode --unsafe | pino-pretty",
+    "start:prod": "yarn build && node dist/bundler-relayer/src/index.js --txMode base| pino-pretty",
     "test": "vitest --config ./vitest.config.ts",
     "lint": "echo no lint",
     "lint:fix": "echo no lint:fix",
