+++ packages/bundler-builder/src/bundle/index.ts
@@ -1,2 +1,3 @@
-export { BundleManager } from './BundleManager.js'
-export { BundleProcessor } from './BundleProcessor.js'
+export * from './bundle-manager.js'
+export * from './bundle-builder.js'
+export * from './bundle-processor.js'
